contact_tracing_service/requirements.txt
grpcio
pymongo
confluent-kafka
protobuf
grpcio-tools
prometheus_client
